# contracts/openapi/install-v1.yaml
openapi: 3.1.0
info:
  title: APP Installation API
  version: 0.4.0-dev
  description: |
    首次安装状态 API。该 scope 与管理端、客户端 API 隔离。
servers:
  - url: /
    description: Local runtime
tags:
  - name: Installation
paths:
  /api/system/install/v1/status:
    get:
      tags: [Installation]
      operationId: getInstallationStatus
      summary: 获取无敏感字段的安装状态
      responses:
        '200':
          description: 安装状态读取成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationStatusEnvelope'
        '500':
          description: 安装状态不可读
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/system/install/v1/plan:
    post:
      tags: [Installation]
      operationId: createInstallationPlan
      summary: 生成 UI 与配置写入的权限预检计划
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
      responses:
        '200':
          description: 预检计划生成成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationPlanEnvelope'
        '400':
          description: 选择或路径预检参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/check/database:
    post:
      tags: [Installation]
      operationId: checkInstallationDatabase
      summary: 测试 MySQL 或 PostgreSQL 连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseConnection'
      responses:
        '200':
          description: 返回不含连接材料的可达性结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyCheckEnvelope'
        '400':
          description: 连接参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/check/redis:
    post:
      tags: [Installation]
      operationId: checkInstallationRedis
      summary: 测试 Redis 连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedisConnection'
      responses:
        '200':
          description: 返回不含连接材料的可达性结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyCheckEnvelope'
        '400':
          description: 连接参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/apply:
    post:
      tags: [Installation]
      operationId: applyInstallation
      summary: 执行已确认的首次安装事务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRequest'
      responses:
        '200':
          description: 安装事务完成
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyResultEnvelope'
        '202':
          description: 安装事务已进入后台执行
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyJobEnvelope'
        '400':
          description: 安装请求无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '409':
          description: 已完成安装或另一个安装事务正在运行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '422':
          description: 权限或依赖预检未通过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '500':
          description: 安装事务失败且已执行补偿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/progress/{id}:
    get:
      tags: [Installation]
      operationId: getInstallationProgress
      summary: 查询后台安装事务进度
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 80
      responses:
        '200':
          description: 安装进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyJobEnvelope'
        '404':
          description: 安装任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/retry/{id}:
    post:
      tags: [Installation]
      operationId: retryInstallation
      summary: 使用新凭据重试失败的安装事务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 80
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRequest'
      responses:
        '202':
          description: 重试任务已进入后台执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyJobEnvelope'
        '400':
          description: 重试请求无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '404':
          description: 安装任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '409':
          description: 安装已完成或已有任务运行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
  /api/system/install/v1/rollback/{id}:
    post:
      tags: [Installation]
      operationId: rollbackInstallation
      summary: 显式确认并回滚失败的安装事务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 80
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackRequest'
      responses:
        '200':
          description: 回滚完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RollbackResultEnvelope'
        '400':
          description: 缺少显式回滚确认或请求无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '404':
          description: 安装任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '409':
          description: 回滚不可用或已有安装事务运行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyErrorEnvelope'
        '503':
          description: 安装服务未接通
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallServiceUnavailableEnvelope'
components:
  headers:
    RequestId:
      description: 请求关联标识
      schema:
        type: string
  schemas:
    InstallationStatusEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: success
        data:
          $ref: '#/components/schemas/InstallationStatus'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    InstallationStatus:
      type: object
      required: [state, installed, schemaVersion, installerVersion]
      properties:
        state:
          type: string
          enum: [pristine, ui_prepared, installing, installed, inconsistent]
        installed:
          type: boolean
        schemaVersion:
          type: integer
          minimum: 1
        installerVersion:
          type: string
        selectedUi:
          type: string
          enum: [antd, ele, naive]
          readOnly: true
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
        installedAt:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 50000
        message:
          type: string
          const: installation state unavailable
        data:
          type: 'null'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
    PlanRequest:
      type: object
      required: [mode]
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
    InstallationPlanEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: success
        data:
          $ref: '#/components/schemas/InstallationPlan'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    InstallationPlan:
      type: object
      required: [selectedUi, mode, canCleanup, canBuild, canWriteEnv, requiresRestart, entries, reasons]
      properties:
        selectedUi:
          type: string
          enum: [antd, ele, naive]
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
        canCleanup:
          type: boolean
          description: 已由命令行初始化完成的模板暂存状态，保留以兼容安装计划摘要。
        canBuild:
          type: boolean
        canWriteEnv:
          type: boolean
        requiresRestart:
          type: boolean
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PlanEntry'
        reasons:
          type: array
          items:
            type: string
    PlanEntry:
      type: object
      required: [path, action, permission]
      properties:
        path:
          type: string
        action:
          type: string
          enum: [keep, write]
        permission:
          $ref: '#/components/schemas/PathPermission'
    PathPermission:
      type: object
      required: [canRead, canWrite, canCreate, canRename, canDelete, reasons]
      properties:
        canRead:
          type: boolean
        canWrite:
          type: boolean
        canCreate:
          type: boolean
        canRename:
          type: boolean
        canDelete:
          type: boolean
        reasons:
          type: array
          items:
            type: string
    PlanErrorEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 10000
        message:
          type: string
          const: invalid installation plan
        data:
          type: 'null'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    InstallServiceUnavailableEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 40001
        message:
          type: string
          const: installation service unavailable
        data:
          type: 'null'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    DatabaseConnection:
      type: object
      required: [driver, mode]
      additionalProperties: false
      properties:
        driver:
          type: string
          enum: [mysql, postgres]
        mode:
          type: string
          enum: [single, read_write, cluster_endpoint]
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        database:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        dsn:
          type: string
          writeOnly: true
        primaryDsn:
          type: string
          writeOnly: true
        replicaDsns:
          type: array
          writeOnly: true
          items:
            type: string
        tlsMode:
          type: string
    RedisConnection:
      type: object
      required: [mode]
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [single, sentinel, cluster]
        addr:
          type: string
        addrs:
          type: array
          items:
            type: string
        masterName:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        db:
          type: integer
          minimum: 0
    DependencyCheckEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: success
        data:
          $ref: '#/components/schemas/DependencyCheck'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    DependencyCheck:
      type: object
      required: [kind, mode, ok, reason]
      properties:
        kind:
          type: string
          enum: [database, redis]
        driver:
          type: string
          enum: [mysql, postgres]
        mode:
          type: string
        ok:
          type: boolean
        reason:
          type: string
          enum: [reachable, connection_failed]
        latencyMs:
          type: integer
          minimum: 0
    ConnectionErrorEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 10000
        message:
          type: string
          enum: [invalid database connection, invalid redis connection]
        data:
          type: 'null'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApplyRequest:
      type: object
      additionalProperties: false
      required: [mode, database, redis, admin]
      properties:
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
        localeMode:
          type: string
          enum: [single, multi]
          default: single
          description: 语言模式；省略时使用 single
        locale:
          type: string
          enum: [zh-CN, en-US]
          default: zh-CN
          description: 安装后默认语言；省略时按浏览器 Accept-Language 建议
        database:
          $ref: '#/components/schemas/DatabaseConnection'
        redis:
          $ref: '#/components/schemas/RedisConnection'
        admin:
          $ref: '#/components/schemas/AdminAccount'
    AdminAccount:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
    ApplyResultEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: success
        data:
          $ref: '#/components/schemas/ApplyResult'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApplyJobEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          enum: [accepted, success]
        data:
          $ref: '#/components/schemas/ApplyJob'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApplyJob:
      type: object
      required: [id, state, mode, currentStep, progress, steps, canRetry, canRollback, lastUpdated]
      properties:
        id:
          type: string
        state:
          type: string
          enum: [running, completed, failed]
        selectedUi:
          type: string
          enum: [antd, ele, naive]
          readOnly: true
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
        currentStep:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ApplyStep'
        installedAt:
          type: string
          format: date-time
        errorCode:
          type: integer
        errorKey:
          type: string
        canRetry:
          type: boolean
        canRollback:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
    RollbackRequest:
      type: object
      required: [confirmRollback]
      additionalProperties: false
      properties:
        confirmRollback:
          type: boolean
          const: true
          description: 必须由操作者明确确认回滚
    RollbackResult:
      type: object
      required: [jobId, state, currentStep, rolledBack, canRetry, lastUpdated]
      additionalProperties: false
      properties:
        jobId:
          type: string
        state:
          type: string
          enum: [failed]
        currentStep:
          type: string
          const: rolled_back
        rolledBack:
          type: boolean
          const: true
        canRetry:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
    RollbackResultEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: success
        data:
          $ref: '#/components/schemas/RollbackResult'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApplyResult:
      type: object
      required: [state, selectedUi, mode, installedAt, steps]
      properties:
        state:
          type: string
          const: installed
        selectedUi:
          type: string
          enum: [antd, ele, naive]
        mode:
          type: string
          enum: [embedded, standalone, api_only, dev]
        installedAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ApplyStep'
    ApplyStep:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          enum: [plan, database, redis, schema, assets, identity, environment, lock]
        status:
          type: string
          const: completed
    ApplyErrorEnvelope:
      type: object
      required: [code, message, data, traceId, meta]
      properties:
        code:
          type: integer
          enum: [10000, 10001, 10006, 10007, 10008, 10009, 30000, 50000]
        message:
          type: string
        data:
          type: 'null'
        traceId:
          type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'